move LayerScale
