+++ tests/test_cnn/test_transformer.py
@@ -7,7 +7,7 @@ from mmengine.model import ModuleList
 
 from mmcv.cnn.bricks.drop import DropPath
 from mmcv.cnn.bricks.transformer import (FFN, AdaptivePadding,
-                                         BaseTransformerLayer, LayerScale,
+                                         BaseTransformerLayer,
                                          MultiheadAttention, PatchEmbed,
                                          PatchMerging,
                                          TransformerLayerSequence)
@@ -558,45 +558,6 @@ def test_ffn():
     input_tensor = torch.rand(2, 20, 256)
     input_tensor_nbc = input_tensor.transpose(0, 1)
     assert torch.allclose(ffn(input_tensor).sum(), ffn(input_tensor_nbc).sum())
-
-
-def test_layer_scale():
-    with pytest.raises(AssertionError):
-        cfg = dict(
-            dim=10,
-            data_format='BNC',
-        )
-        LayerScale(**cfg)
-
-    # test init
-    cfg = dict(dim=10)
-    ls = LayerScale(**cfg)
-    assert torch.equal(ls.weight, torch.ones(10, requires_grad=True) * 1e-5)
-
-    # test forward
-    # test channels_last
-    cfg = dict(dim=256, inplace=False, data_format='channels_last')
-    ls_channels_last = LayerScale(**cfg)
-    x = torch.randn((4, 49, 256))
-    out = ls_channels_last(x)
-    assert tuple(out.size()) == (4, 49, 256)
-    assert torch.equal(x * 1e-5, out)
-
-    # test channels_first
-    cfg = dict(dim=256, inplace=False, data_format='channels_first')
-    ls_channels_first = LayerScale(**cfg)
-    x = torch.randn((4, 256, 7, 7))
-    out = ls_channels_first(x)
-    assert tuple(out.size()) == (4, 256, 7, 7)
-    assert torch.equal(x * 1e-5, out)
-
-    # test inplace True
-    cfg = dict(dim=256, inplace=True, data_format='channels_first')
-    ls_channels_first = LayerScale(**cfg)
-    x = torch.randn((4, 256, 7, 7))
-    out = ls_channels_first(x)
-    assert tuple(out.size()) == (4, 256, 7, 7)
-    assert x is out
 
 
 @pytest.mark.skipif(not torch.cuda.is_available(), reason='Cuda not available')
